Revert "Auth Redirect to Login and Auth Config (#48)"
This reverts commit 81c6235f2af128a60a30e83a41559fa0ec394a31.
