+++ app/src/App.tsx
@@ -5,16 +5,16 @@ import {AirTableService, FirebaseService, Prize, Task} from './Api';
 import './App.css';
 import Home from "./pages/Home";
 
-import { AuthProvider } from "./AuthContext";
-
 import {
     getAuth,
     User,
+    signInWithPopup,
+    GoogleAuthProvider,
+    sendSignInLinkToEmail,
+    isSignInWithEmailLink,
+    signInWithEmailLink,
+    signOut,
 } from "firebase/auth";
-
-
-import { getAnalytics } from 'firebase/analytics';
-
 
 import Login from "./pages/Login";
 import RaffleListView from "./pages/RaffleListView";
@@ -48,7 +48,6 @@ const firebaseConfig = {
 
 const firebaseApp = initializeApp(firebaseConfig);
 const auth = getAuth(firebaseApp);
-const analytics = getAnalytics(firebaseApp);
 const firebaseService = new FirebaseService();
 const airtableService = new AirTableService();
 
@@ -66,7 +65,6 @@ export type UserData = {
     raffles_entered?: {},
     tickets_remaining?: number
 }
-
 
 function App() {
     const [user, setUser] = useState<User | null>(null);
@@ -86,16 +84,44 @@ function App() {
       return allTasks;
     };
 
+    const signInWithGoogle = async () => {
+        const provider = new GoogleAuthProvider();
+        try {
+            await signInWithPopup(auth, provider);
+        } catch (error: any) {
+            setError(error.message);
+        }
+    };
+
+    const sendSignInEmail = async (email: string) => {
+        const actionCodeSettings = {
+            url: window.location.href,
+            handleCodeInApp: true,
+        };
+        try {
+            await sendSignInLinkToEmail(auth, email, actionCodeSettings);
+            window.localStorage.setItem("emailForSignIn", email);
+            setEmail("");
+            setError("Email sent. Please check your inbox.");
+        } catch (error: any) {
+            setError(error.message);
+        }
+    };
+
+    const signInWithEmail = async (email: string, emailLink: string) => {
+        try {
+            await signInWithEmailLink(auth, email, emailLink);
+        } catch (error: any) {
+            setError(error.message);
+        }
+    };
+
     interface UserProps {
         user: User | null;
     }
 
     function HomePage(props: UserProps) {
         const { user } = props;
-        //   const navigate = useNavigate();
-        if (!user && window.location.pathname !== "/login") {
-            window.location.pathname = "/login";
-        }
         return (
             <div>
                 {user ? (
@@ -103,7 +129,7 @@ function App() {
                 ) : (
                     <>
                         <h3>Sign in to get started!</h3>
-                        {/* <button onClick={signInWithGoogle}>Sign in with Google</button> */}
+                        <button onClick={signInWithGoogle}>Sign in with Google</button>
                         <form onSubmit={handleSubmit}>
                             <input
                                 type="email"
@@ -130,13 +156,13 @@ function App() {
         });
 
         // @TODO https://firebase.google.com/docs/auth/web/email-link-auth?authuser=2&hl=en
-        // if (isSignInWithEmailLink(auth, window.location.href)) {
-        //     const email = window.localStorage.getItem("emailForSignIn");
-        //     if (email) {
-        //         signInWithEmail(email, window.location.href);
-        //         window.localStorage.removeItem("emailForSignIn");
-        //     }
-        // }
+        if (isSignInWithEmailLink(auth, window.location.href)) {
+            const email = window.localStorage.getItem("emailForSignIn");
+            if (email) {
+                signInWithEmail(email, window.location.href);
+                window.localStorage.removeItem("emailForSignIn");
+            }
+        }
         if (user) {
             // Fetch user data
             const fetchUserData = async () => {
@@ -207,7 +233,7 @@ function App() {
 
     const handleSubmit = (event: React.FormEvent) => {
         event.preventDefault();
-        // sendSignInEmail(email);
+        sendSignInEmail(email);
     };
 
     const router = createBrowserRouter([
@@ -235,9 +261,7 @@ function App() {
 
     return (
         <div className="App">
-            <AuthProvider appContext={firebaseApp}>
-                <RouterProvider router={router} />
-            </AuthProvider>
+            <RouterProvider router={router} />
         </div>
     );
 }
